+++ extensions/united-nations/CHANGELOG.md
@@ -1,5 +1,10 @@
 # United Nations Extension
 
+## [Maintenance] - 2025-06-04
+
+- Drop unexpected LaTeX content
+- Bump all dependencies to the latest
+
 ## [Bugfix] - 2025-06-03
 
 - Replace broken UN documents GA RSS feed source